requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
psycopg2-binary
pydantic>=2.0
python-jose[cryptography]
passlib[bcrypt]

# Testing
pytest
httpx
faker